+++ tw-tasks-core/src/main/java/com/transferwise/tasks/ITasksService.java
@@ -150,5 +150,4 @@ public interface ITasksService {
   enum TasksProcessingState {
     STARTED, STOPPED, STOP_IN_PROGRESS
   }
-
 }