+++ tw-tasks-core/src/main/java/com/transferwise/tasks/domain/TaskStatus.java
@@ -8,5 +8,6 @@ public enum TaskStatus {
   DONE,
   ERROR, // Generates alerts
   FAILED, // ERROR is acked, alerts are off
+  CANCELLED, // User requested cancellation, no alerts
   UNKNOWN // For metrics, if getting task status is too expensive
 }