+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.50.0 - 2025/02/12
+
+### Changed
+
+- Added support to cancel tasks in waiting state.
+
+
 ## 1.49.0 - 2025/01/08
 
 ### Changed
@@ -18,7 +25,7 @@ It is worth keeping an eye on:
   - changes to assignors used, log `Successfully synced group in generation Generation`
   - on assignment strategy failures on consumers in prod and [consumer state](https://dashboards.tw.ee/d/f7094f30-a509-4592-aced-37584a70132a/kafka-consumer-groups-and-lag-details-kminion?orgId=1&refresh=30s&viewPanel=14).
 
-If you use `com.wise.kafka.assignors.CanaryAwareRangeAssignor`, consider setting this config: 
+If you use `com.wise.kafka.assignors.CanaryAwareRangeAssignor`, consider setting this config:
 ```
 spring.kafka.consumer.properties.partition.assignment.strategy:
 com.wise.kafka.assignors.CanaryAwareRangeAssignor, org.apache.kafka.clients.consumer.RangeAssignor, org.apache.kafka.clients.consumer.CooperativeStickyAssignor